+++ pkg/compactor/partition_compaction_grouper.go
@@ -45,6 +45,7 @@ type PartitionCompactionGrouper struct {
 	blockFilesConcurrency    int
 	blocksFetchConcurrency   int
 	compactionConcurrency    int
+	totalGroupsPlanned       int
 
 	doRandomPick bool
 
@@ -114,6 +115,11 @@ func NewPartitionCompactionGrouper(
 
 // Groups function modified from https://github.com/cortexproject/cortex/pull/2616
 func (g *PartitionCompactionGrouper) Groups(blocks map[ulid.ULID]*metadata.Meta) (res []*compact.Group, err error) {
+	remainingConcurrency := g.compactionConcurrency - g.totalGroupsPlanned
+	if remainingConcurrency <= 0 {
+		return nil, nil
+	}
+
 	// Check if this compactor is on the subring.
 	// If the compactor is not on the subring when using the userID as a identifier
 	// no plans generated below will be owned by the compactor so we can just return an empty array
@@ -140,7 +146,8 @@ func (g *PartitionCompactionGrouper) Groups(blocks map[ulid.ULID]*metadata.Meta)
 		return nil, errors.Wrap(err, "unable to generate compaction jobs")
 	}
 
-	pickedPartitionCompactionJobs := g.pickPartitionCompactionJob(partitionCompactionJobs)
+	pickedPartitionCompactionJobs := g.pickPartitionCompactionJob(partitionCompactionJobs, remainingConcurrency)
+	g.totalGroupsPlanned += len(pickedPartitionCompactionJobs)
 
 	return pickedPartitionCompactionJobs, nil
 }
@@ -622,7 +629,7 @@ func (g *PartitionCompactionGrouper) handleEmptyPartition(partitionedGroupInfo *
 	return nil
 }
 
-func (g *PartitionCompactionGrouper) pickPartitionCompactionJob(partitionCompactionJobs []*blocksGroupWithPartition) []*compact.Group {
+func (g *PartitionCompactionGrouper) pickPartitionCompactionJob(partitionCompactionJobs []*blocksGroupWithPartition, remainingConcurrency int) []*compact.Group {
 	var outGroups []*compact.Group
 	for _, partitionedGroup := range partitionCompactionJobs {
 		groupHash := partitionedGroup.groupHash
@@ -697,7 +704,7 @@ func (g *PartitionCompactionGrouper) pickPartitionCompactionJob(partitionCompact
 
 		outGroups = append(outGroups, thanosGroup)
 		level.Debug(partitionedGroupLogger).Log("msg", "added partition to compaction groups")
-		if len(outGroups) >= g.compactionConcurrency {
+		if len(outGroups) >= remainingConcurrency {
 			break
 		}
 	}